app/requirements.txt
openai
streamlit
psycopg2-binary
numpy==1.26.4
pandas==2.2.1
python-dotenv